Made a few minor changes to improve UI, committing so that I can pull in master
