+++ Assets/Scripts/CardHandlerScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ebca44ad77350dc47917aac8caf90cb8
+timeCreated: 1477943619
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
