+++ Assets/Scripts/PlayerSelectionController.cs
@@ -10,7 +10,7 @@ public class PlayerSelectionController : MonoBehaviour {
 
     int playersCount = 1;
     int currentPlayer = 1;
-    int tempPlayerCount = 0;
+    //int tempPlayerCount = 0;
 
     string playerCountKeyString = "";
     string currentPlayerTribeKeyString = "";
@@ -21,6 +21,14 @@ public class PlayerSelectionController : MonoBehaviour {
     string currentPlayerNotorietyKeyString = "";
     
     // Use this for initialization
+    void Awake()
+    {
+        if (FindObjectsOfType<PlayerSelectionController>().Length > 1)
+        {
+            DestroyImmediate(this);
+        }
+    }
+
 	void Start () 
     {
         // Makes the object persistent between scenes
@@ -45,7 +53,7 @@ public class PlayerSelectionController : MonoBehaviour {
         currentPlayerSurvivalKeyString = "SurvivalStat" + currentPlayer.ToString();
         currentPlayerNotorietyKeyString = "NotorietyStat" + currentPlayer.ToString();
 
-        print("player switch initiated");
+        print("Player " + currentPlayer.ToString() + "'s turn");
     }
 
     public string GetCurrentTribeTypeKey()
@@ -83,13 +91,18 @@ public class PlayerSelectionController : MonoBehaviour {
         playersCount = numberOfPlayers;
     }
 
-    public int GetTempPlayerCount()
+    public int GetCurrentPlayerNumber()
     {
-        return tempPlayerCount;
+        return currentPlayer;
     }
 
-    public void SetTempPlayerCount(int newTempPlayerCount)
-    {
-        tempPlayerCount = newTempPlayerCount;
-    }
+    //public int GetTempPlayerCount()
+    //{
+    //    return tempPlayerCount;
+    //}
+
+    //public void SetTempPlayerCount(int newTempPlayerCount)
+    //{
+    //    tempPlayerCount = newTempPlayerCount;
+    //}
 }
