+++ Assets/Scripts/PlayerSetupButtonScript.cs
@@ -1,0 +1,273 @@
+﻿using UnityEngine;
+using System.Collections;
+using UnityEngine.SceneManagement;
+using UnityEngine.UI;
+
+public class PlayerSetupButtonScript : MonoBehaviour {
+
+    public string sceneName;
+
+    // Buttons passed in
+    public PlayerSelectionController playerSelectionController;
+
+    // The number of players in the game
+    int players;
+
+    // The input field for the player's tribe name
+    public InputField tribeNameInputField;
+
+    // The text field that shows which player's turn it is
+    public Text playerNumberText;
+
+    // Tribe type enumeration
+    public enum TribeType
+    {
+        Type0,
+        Type1,
+        Type2,
+        Type3
+    }
+
+    // The player's decided tribe type
+    public TribeType tribeType;
+
+    // Tribe Type Buttons
+    public Button okbutton;
+    public Button tribe0Button;
+    public Button tribe1Button;
+    public Button tribe2Button;
+    public Button tribe3Button;
+
+    // Tribe Type Text
+    public Text tribe0FlavorText;
+    public Text tribe1FlavorText;
+    public Text tribe2FlavorText;
+    public Text tribe3FlavorText;
+
+    public Image finalDecisionPanel;
+
+    // Sliders for stats
+    public Slider strengthSlider;
+    public Slider agilitySlider;
+    public Slider trustSlider;
+    public Slider survivalSlider;
+    public Slider notorietySlider;
+
+    // Stat Values
+    public int strengthStat;
+    public int agilityStat;
+    public int trustStat;
+    public int survivalStat;
+    public int notorietyStat;
+
+    public void Start()
+    {
+        playerSelectionController = GameObject.FindObjectOfType<PlayerSelectionController>();
+
+        players = PlayerPrefs.GetInt("NumberOfPlayers");
+
+        print("Number of Players: " + players.ToString());
+
+        ResetStats();
+    }
+	
+	// Update is called once per frame
+	void Update () 
+    {
+	
+	}
+
+    public void ChangeScene()
+    {
+        // Loops the player setup scene for the number of players selected
+        if (SceneManager.GetActiveScene().name == "PlayerSetup")
+        {
+            if (playerSelectionController.GetCurrentPlayerNumber() < players)
+            {
+                sceneName = "PlayerSetup";
+                playerSelectionController.PlayerSwitch(playerSelectionController.GetCurrentPlayerNumber() + 1);
+
+                UpdatePlayerNumber(playerSelectionController.GetCurrentPlayerNumber());
+
+                ResetStats();
+            }
+            else
+            {
+                sceneName = "Gameplay";
+            }
+        }
+
+        SceneManager.LoadScene(sceneName);
+    }
+
+    public void ActivateOK()
+    {
+        if (okbutton.gameObject.activeInHierarchy == false)
+        {
+            okbutton.gameObject.SetActive(true);
+        }
+    }
+
+    public void Tribe0Flavor()
+    {
+        tribeType = TribeType.Type0;
+
+        // This is where the selected value of the tribe is stored between scenes
+        PlayerPrefs.SetString(playerSelectionController.GetCurrentTribeTypeKey(), tribeType.ToString());
+
+        tribe0FlavorText.gameObject.SetActive(true);
+        tribe1FlavorText.gameObject.SetActive(false);
+        tribe2FlavorText.gameObject.SetActive(false);
+        tribe3FlavorText.gameObject.SetActive(false);
+
+        ActivateOK();
+    }
+
+    public void Tribe1Flavor()
+    {
+        tribeType = TribeType.Type1;
+
+        // This is where the selected value of the tribe is stored between scenes
+        PlayerPrefs.SetString(playerSelectionController.GetCurrentTribeTypeKey(), tribeType.ToString());
+
+        tribe0FlavorText.gameObject.SetActive(false);
+        tribe1FlavorText.gameObject.SetActive(true);
+        tribe2FlavorText.gameObject.SetActive(false);
+        tribe3FlavorText.gameObject.SetActive(false);
+
+        ActivateOK();
+    }
+
+    public void Tribe2Flavor()
+    {
+        tribeType = TribeType.Type2;
+
+        // This is where the selected value of the tribe is stored between scenes
+        PlayerPrefs.SetString(playerSelectionController.GetCurrentTribeTypeKey(), tribeType.ToString());
+
+        tribe0FlavorText.gameObject.SetActive(false);
+        tribe1FlavorText.gameObject.SetActive(false);
+        tribe2FlavorText.gameObject.SetActive(true);
+        tribe3FlavorText.gameObject.SetActive(false);
+
+        ActivateOK();
+    }
+
+    public void Tribe3Flavor()
+    {
+        tribeType = TribeType.Type3;
+
+        // This is where the selected value of the tribe is stored between scenes
+        PlayerPrefs.SetString(playerSelectionController.GetCurrentTribeTypeKey(), tribeType.ToString());
+
+        tribe0FlavorText.gameObject.SetActive(false);
+        tribe1FlavorText.gameObject.SetActive(false);
+        tribe2FlavorText.gameObject.SetActive(false);
+        tribe3FlavorText.gameObject.SetActive(true);
+
+        ActivateOK();
+    }
+
+    public void StrengthUpdate()
+    {
+        strengthStat = (int)strengthSlider.value;
+
+        // This is where the Strength stat is stored between scenes
+        PlayerPrefs.SetInt(playerSelectionController.GetCurrentStrengthStatKey(), strengthStat);
+    }
+
+    public void AgilityUpdate()
+    {
+        agilityStat = (int)agilitySlider.value;
+
+        // This is where the Agility stat is stored between scenes
+        PlayerPrefs.SetInt(playerSelectionController.GetCurrentAgilityStatKey(), agilityStat);
+    }
+
+    public void TrustUpdate()
+    {
+        trustStat = (int)trustSlider.value;
+
+        // This is where the Trust stat is stored between scenes
+        PlayerPrefs.SetInt(playerSelectionController.GetCurrentTrustStatKey(), trustStat);
+    }
+
+    public void SurvivalUpdate()
+    {
+        survivalStat = (int)survivalSlider.value;
+
+        // This is where the Survival stat is stored between scenes
+        PlayerPrefs.SetInt(playerSelectionController.GetCurrentSurvivalStatKey(), survivalStat);
+    }
+
+    public void NotorietyUpdate()
+    {
+        notorietyStat = (int)notorietySlider.value;
+
+        // This is where the Notoriety stat is stored between scenes
+        PlayerPrefs.SetInt(playerSelectionController.GetCurrentNotorietyStatKey(), notorietyStat);
+    }
+
+    public void OnOKClick()
+    {
+        tribe0Button.gameObject.SetActive(false);
+        tribe1Button.gameObject.SetActive(false);
+        tribe2Button.gameObject.SetActive(false);
+        tribe3Button.gameObject.SetActive(false);
+
+        strengthSlider.gameObject.SetActive(false);
+        agilitySlider.gameObject.SetActive(false);
+        trustSlider.gameObject.SetActive(false);
+        survivalSlider.gameObject.SetActive(false);
+        notorietySlider.gameObject.SetActive(false);
+
+        tribeNameInputField.gameObject.SetActive(false);
+
+        okbutton.gameObject.SetActive(false);
+
+        finalDecisionPanel.gameObject.SetActive(true);
+    }
+
+    public void OnNoClick()
+    {
+        tribe0Button.gameObject.SetActive(true);
+        tribe1Button.gameObject.SetActive(true);
+        tribe2Button.gameObject.SetActive(true);
+        tribe3Button.gameObject.SetActive(true);
+
+        strengthSlider.gameObject.SetActive(true);
+        agilitySlider.gameObject.SetActive(true);
+        trustSlider.gameObject.SetActive(true);
+        survivalSlider.gameObject.SetActive(true);
+        notorietySlider.gameObject.SetActive(true);
+
+        tribeNameInputField.gameObject.SetActive(true);
+
+        okbutton.gameObject.SetActive(true);
+
+        finalDecisionPanel.gameObject.SetActive(false);
+    }
+
+    public void UpdatePlayerNumber(int newPlayerNumber)
+    {
+        playerNumberText.text = "Player " + newPlayerNumber.ToString() + "'s turn";
+    }
+
+    private void ResetStats()
+    {
+        tribeType = TribeType.Type0;
+        strengthStat = 0;
+        agilityStat = 0;
+        trustStat = 0;
+        survivalStat = 0;
+        notorietyStat = 0;
+
+        // Initializing all stats and values to their default states
+        PlayerPrefs.SetString(playerSelectionController.GetCurrentTribeTypeKey(), tribeType.ToString());
+        PlayerPrefs.SetInt(playerSelectionController.GetCurrentStrengthStatKey(), strengthStat);
+        PlayerPrefs.SetInt(playerSelectionController.GetCurrentAgilityStatKey(), agilityStat);
+        PlayerPrefs.SetInt(playerSelectionController.GetCurrentTrustStatKey(), trustStat);
+        PlayerPrefs.SetInt(playerSelectionController.GetCurrentSurvivalStatKey(), survivalStat);
+        PlayerPrefs.SetInt(playerSelectionController.GetCurrentNotorietyStatKey(), notorietyStat);
+    }
+}
