+++ Assets/Scripts/PlayerSetupButtonScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 4cfe1af18661a7548a66399ecad49eab
+timeCreated: 1478080543
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
