+++ Assets/Scripts/SliderScript.cs
@@ -4,7 +4,6 @@ using UnityEngine.UI;
 
 public class SliderScript : MonoBehaviour {
 
-    public Button playerSlideButton;
     public Slider mainSlider;
     public Text sliderButtonText;
 
@@ -13,7 +12,6 @@ public class SliderScript : MonoBehaviour {
     void Start()
     {
         sliderButtonText.text = "Player Count: (2 -4 ) ";
-        Debug.Log("textDisplayed");
     }
 
     void Update()
@@ -22,7 +20,6 @@ public class SliderScript : MonoBehaviour {
         {
             players = (int)mainSlider.value;
             sliderButtonText.text = "Player Count : (2 -4)  " + players;
-            Debug.Log("Slider Updated");
         }
         
     }
