+++ Assets/Scripts/StatSliderScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 8d7dafe8bc72d224d86839940f5817e2
+timeCreated: 1478117651
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
