+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 5.3.2f1
+m_EditorVersion: 5.4.1f1
 m_StandardAssetsVersion: 0
